trapframe 2.1.2

+++ xv6-riscv/kernel/trap.c
@@ -128,7 +128,7 @@ usertrapret(void)
   uint64 trampoline_userret = TRAMPOLINE + (userret - trampoline);
   
   // CSE 536: (Task 2.1.2) - Pass the address of the current thread's TRAPFRAME in s11
-  p->thread->trapframe->s11=p->thread->trapframe;
+  p->thread->trapframe->s11=(uint64)p->thread->trapframe;
   ((void (*)(uint64))trampoline_userret)(satp);
 }
